requirements.txt
Flask==3.0.3
google-generativeai==0.5.2
